Day21 長輩圖生成器/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pillow==10.4.0
python-multipart==0.0.9
pydantic==2.9.2
google-genai==0.3.0
openai==1.100.0
aiofiles==24.1.0
requests==2.32.3
onnxruntime==1.18.1
qrcode==7.4.2
loguru==0.7.2
pytest==8.3.2
starlette==0.38.5
httpx==0.27.2
